Fix minor UI issues

+++ site/src/util/filetree.ts
@@ -7,7 +7,38 @@ export type FileTree = {
   [key: string]: FileTree | string
 }
 
-export const setFile = (
+export const createFile = (
+  path: string,
+  fileTree: FileTree,
+  value: string,
+): FileTree => {
+  if (existsFile(path, fileTree)) {
+    throw new Error(`File ${path} already exists`)
+  }
+  const pathError = validatePath(path, fileTree)
+  if (pathError) {
+    throw new Error(pathError)
+  }
+
+  return set(fileTree, path.split("/"), value)
+}
+
+export const validatePath = (
+  path: string,
+  fileTree: FileTree,
+): string | undefined => {
+  const paths = path.split("/")
+  paths.pop() // The last item is the filename
+  for (let i = 0; i <= paths.length; i++) {
+    const path = paths.slice(0, i + 1)
+    const pathStr = path.join("/")
+    if (existsFile(pathStr, fileTree) && !isFolder(pathStr, fileTree)) {
+      return `Invalid path. The path ${path} is not a folder`
+    }
+  }
+}
+
+export const updateFile = (
   path: string,
   content: FileTree | string,
   fileTree: FileTree,
@@ -31,8 +62,11 @@ export const moveFile = (
   fileTree: FileTree,
 ) => {
   const content = getFileContent(currentPath, fileTree)
+  if (typeof content !== "string") {
+    throw new Error("Move folders is not allowed")
+  }
   fileTree = removeFile(currentPath, fileTree)
-  fileTree = setFile(newPath, content, fileTree)
+  fileTree = createFile(newPath, fileTree, content)
   return fileTree
 }
